+++ test/lib/validators.js
@@ -82,6 +82,13 @@ const createValidateLine = (cfg) => {
 
 	const validateLine = (val, line, name = 'line') => {
 		defaultValidators.line(val, line, name)
+
+		const msg = name + '.fahrtNr must be '
+		if (line.fahrtNr !== null) {
+			a.strictEqual(typeof line.fahrtNr, 'string', msg + 'a string')
+			a.ok(line.fahrtNr, msg + ' be empty')
+		}
+
 		a.ok(validLineModes.includes(line.mode), name + '.mode is invalid')
 	}
 	return validateLine