+++ docs/changelog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## `3.2.0`
+
+- #79 [CapMetro/CMTA profile](p/cmta) – Thanks @nickturskyi!
+
+## `3.1.2`
+
+- f796337 handle warnings without schedule `sDate`/`eDate`/`lModDate` 🐛
+
+## `3.1.1`
+
+- 39cc2f3 fix install on Windows 🐛
+
+## `3.1.0`
+
+- 9257d3a parse `line.fahrtNr`
+
 ## `3.0.0`
 
 This version is not fully backwords-compatible. Check out [the migration guide](migrating-to-3.md).